Add and use helper to only match when not in tag attribute values
This is far from being perfect but it does prevent transforming things within tag attributes values. Unfortunately it also prevents transforming things like the colon in: `<p>Hello color: red" attr ></p>` because the right part up to the closing </p> looks like what would be within a html tag - But it's better not to transform those than to transform what could be part of an html tag and mess up inline styles.